+++ frontend/src/components/authentication/AuthenticationTabs.jsx
